refactoring

+++ src/header/Evaluator.hpp
@@ -1,6 +1,8 @@
 #ifndef EVALUATOR_
 #define EVALUATOR_
 
+#include <string.h>
+
 #include "PtrWalker.hpp"
 
 class Evaluator {
@@ -8,7 +10,7 @@ private:
   PtrWalker<char>* byteCode;
   
 public:
-  void eval(char* input, int len);
+  void eval(char* input, size_t len);
 };
 
 #endif
